+++ controllers/program_controller.js
@@ -13,7 +13,7 @@ programs.get('/', async (req, res) => {
       include: [
         {
           model: user,
-          as: 'user'
+          as: 'author'
         }
       ]
     })
@@ -23,21 +23,76 @@ programs.get('/', async (req, res) => {
   }
 })
 
+// FIND PROGRAM BY ID
+programs.get('/:id', async (req, res) => {
+
+  try {
+    const foundProgram = await program.findOne({
+      where: {
+        program_id: req.params.id
+      },
+      include: [
+        {
+          model: user,
+          as: 'author'
+        }
+      ]
+    })
+    res.status(200).json({
+      title: foundProgram.title,
+      author: foundProgram.author,
+      source_code: foundProgram.source_code,
+      binary: foundProgram.binary,
+      description: foundProgram.description
+    })
+  } catch (err) {
+    console.log(err)
+    res.status(500).json(err)
+  }
+})
+
 // CREATE A PROGRAM
-programs.post('/', (req, res) => {
-  res.send('Got a POST request')
+programs.post('/', async (req, res) => {
+
+  if (req.currentUser === null) {
+    return res.status(403).json({ message: 'You must be logged in to post a program' })
+  }
+
+  let { binary, ...rest } = req.body
+
+  console.log(binary)
+  console.log(typeof binary)
+
+  const newProgram = await program.create({
+    ...rest,
+    user_id: req.currentUser.user_id,
+    binary: Buffer.from(binary)
+  })
+
+  res.json(newProgram)
 })
 
 // UPDATE A PROGRAM
 programs.put('/:id', async (req, res) => {
+
+  const foundProgram = await program.findOne({
+    where: {
+      program_id: req.params.id
+    }
+  })
+
+  if (req.currentUser === null || req.currentUser.user_id !== foundProgram.user_id) {
+    return res.status(403).json({ message: 'You must be logged in as the author to edit this program' })
+  }
+
   try {
-    const updatedProgram = await program.update(req.body, {
+    const updatedProgram = await program.update({ binary: Buffer.from(req.body.binary), ...req.body }, {
       where: {
         program_id: req.params.id
       }
     })
     res.status(200).json({
-      message: `Successfully updated ${updatedProgram} program(s)`
+      message: `Successfully updated`
     })
   } catch (err) {
     res.status(500).json(err)
@@ -46,6 +101,11 @@ programs.put('/:id', async (req, res) => {
 
 // DELETE A PROGRAM
 programs.delete('/:id', async (req, res) => {
+
+  if (req.currentUser === null || req.currentUser.user_id !== foundProgram.user_id) {
+    return res.status(403).json({ message: 'You must be logged in as the author to delete this program' })
+  }
+
   try {
     const deletedProgram = await program.destroy({
       where: {
@@ -53,7 +113,7 @@ programs.delete('/:id', async (req, res) => {
       }
     })
     res.status(200).json({
-      message: `Successfully deleted ${deletedProgram} program(s)`
+      message: `Successfully deleted ${deletedProgram.title}`
     })
   } catch (err) {
     res.status(500).json(err)
